refactored #307 - moved forcusmessage into a separate file
