+++ src/Main.js
@@ -8,7 +8,7 @@ import {getBookmark} from './UserProfile';
 import React, { Component } from 'react';
 import MessageList from './MessageList';
 import config, {constant} from './config/default';
-import { default as dist } from './Distance';
+import FocusMessage from './FocusMessage';
 import {
   updateRecentMessage,
   updatePublicProfileDialog,
@@ -88,52 +88,12 @@ class Main extends Component {
     } 
   }
 
-  renderRandomFocusMessages() {
-    const {globalFocusMessages: focusMessages} = this.props.ourland;
-    if(focusMessages != null  && focusMessages.length > 0) {
-      let focusMsgIdx = this.generateFocusMessagesIndex(focusMessages.length);
-      return <div className="focus-message-wrapper">
-          <h4>{focusMessages[focusMsgIdx].title}</h4>
-          <MessageList
-            ref={(messageList) => {this.messageList = messageList;}}
-            eventNumber={100}
-            distance={10}
-            messageIds={focusMessages[focusMsgIdx].messages}
-            hori={true}
-          />
-        </div>
-    }else {
-      return (null);
-    }
-  }
-
-  generateFocusMessagesIndex(length) {
-    // generating a random index for selecting a focus message item randomly
-    // setting focusMsgIdx to state to avoid the value from being changed during re-rendering
-
-    if(!length) return 0;
-    let focusMsgIdx = null;
-    if(this.state.focusMsgIdx == null) {
-      focusMsgIdx = Math.floor((Math.random() * length));
-      this.setState({focusMsgIdx})
-      return focusMsgIdx;
-    } else {
-      return this.state.focusMsgIdx;
-    }
-  }
-
-
   renderMessageFrontPage() {
     let recentMessage = null;
     const { eventNumber, distance, geolocation, eventId, queryMessage, bookmark} = this.state;
     const {open: openRecent} = this.props.recentMessage;
-//    const {focusMessages} = this.props.ourland;
-    const {globalFocusMessages: focusMessages} = this.props.ourland;
-    const { addressBook, classes } = this.props;
-
-    let focusMessage = null
+    const { classes } = this.props;
     let tagStatHtml = this.renderTagStat();
-    let focusMessageHtml = null;
 
     if(queryMessage != null) {
       let message = queryMessage;
@@ -148,74 +108,11 @@ class Main extends Component {
                       </div>;
     }
 
-    // showing focus messages based on users' home address / office address within interested radius
-    if(focusMessages != null  && focusMessages.length > 0 && addressBook != null && addressBook.addresses.length > 0) {
-      // home
-      var homeAddress = addressBook.addresses[0];
-      var homeAddressInterestRadius = homeAddress.distance || 1;
-      // office
-      var officeAddress = addressBook.addresses[1];
-      var officeAddressInterestRadius = officeAddress.distance || 1;
-
-      // generating focus message html array 
-      focusMessageHtml = focusMessages.map((focusMessage, focusMsgIdx) => {
-        let homeAddressDistDiff = null;
-        let officeAddressDistDiff = null;
-
-        if(homeAddress.geolocation != null) {
-          homeAddressDistDiff = dist(focusMessage.geolocation.longitude, focusMessage.geolocation.latitude, homeAddress.geolocation.longitude, homeAddress.geolocation.latitude);
-        }
-
-        if(officeAddress.geolocation) {
-          officeAddressDistDiff = dist(focusMessage.geolocation.longitude, focusMessage.geolocation.latitude, officeAddress.geolocation.longitude, officeAddress.geolocation.latitude);
-        }
-
-        if((homeAddressDistDiff != null && homeAddressDistDiff < homeAddressInterestRadius) || (officeAddressDistDiff != null && officeAddressDistDiff < homeAddressInterestRadius) ) {
-          return (
-            <div key={focusMsgIdx}>
-              <h4>{focusMessages[focusMsgIdx].title}</h4>
-              <MessageList
-                ref={(messageList) => {this.messageList = messageList;}}
-                eventNumber={100}
-                distance={10}
-                messageIds={focusMessages[focusMsgIdx].messages}
-                hori={true}
-              />
-            </div>
-          ) 
-        } else {
-          return null;
-        }
-      });
-
-      // filtering out null cases
-      focusMessageHtml = focusMessageHtml.filter(x => !!x);
-
-      if(focusMessageHtml.length > 0) {
-        // generating focus message
-        // if there are more than one hit, one of them will be selected randomly
-        let focusMsgIdx = this.generateFocusMessagesIndex(focusMessageHtml.length);
-
-        focusMessage = 
-          <div className="focus-message-wrapper">
-            {focusMessageHtml[focusMsgIdx]}
-          </div>
-      } else {
-        // no nearby focusMessage within user's addresses 
-        // generating one randomly from the focus message pool
-        focusMessage = this.renderRandomFocusMessages();
-      } 
-    } else {
-      // no address book for guests
-      // generating one randomly from the focus message pool
-      focusMessage = this.renderRandomFocusMessages();
-    }
-
     return (
       <div className={classes.container}>
         {tagStatHtml}
         {recentMessage}
-        {focusMessage}       
+        <FocusMessage/>      
         <RegionEventDialog
           eventNumber={eventNumber}
           distance={distance}
@@ -240,7 +137,6 @@ const mapStateToProps = (state, ownProps) => {
   return {
     recentMessage : state.recentMessage,
     ourland : state.ourland,
-    addressBook: state.addressBook
   };
 }
